# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Amie API",
    "version": "1.0.0",
    "description": "Amie 애플리케이션의 API 문서"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "개발 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/admin/register": {
      "post": {
        "summary": "관리자 계정 생성",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "name"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "adminPassword"
                  },
                  "name": {
                    "type": "string",
                    "example": "관리자명"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "관리자 계정 생성 성공"
          },
          "400": {
            "description": "잘못된 입력 데이터"
          },
          "409": {
            "description": "중복된 사용자 이름"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/admin/login": {
      "post": {
        "summary": "관리자 로그인",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "adminPassword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/admin": {
      "get": {
        "summary": "모든 관리자 목록 조회",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "관리자 목록"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음 (관리자 아님)"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/admin/{id}": {
      "get": {
        "summary": "특정 관리자 정보 조회",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "관리자 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "관리자 정보"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음 (관리자 아님)"
          },
          "404": {
            "description": "관리자를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "delete": {
        "summary": "관리자 계정 삭제",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "관리자 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "관리자 삭제 성공"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음 (관리자 아님)"
          },
          "404": {
            "description": "관리자를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/admin/{id}/change-password": {
      "put": {
        "summary": "관리자 비밀번호 변경",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "관리자 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "format": "password",
                    "example": "oldPassword"
                  },
                  "newPassword": {
                    "type": "string",
                    "format": "password",
                    "example": "newPassword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "비밀번호 변경 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음 (관리자 아님)"
          },
          "404": {
            "description": "관리자를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
  },
...
}